PdInstance remove useless cast

+++ Source/Pd/PdInstance.cpp
@@ -161,7 +161,7 @@ namespace pd
     void Instance::prepareDSP(const int nins, const int nouts, const double samplerate)
     {
         libpd_set_instance(static_cast<t_pdinstance *>(m_instance));
-        libpd_init_audio((int)nins, (int)nouts, (int)samplerate);
+        libpd_init_audio(nins, nouts, (int)samplerate);
     }
     
     void Instance::startDSP()
